Add index page for GitHub Pages
This is a modified version of the original website, with most of the
information removed and linking to the GitHub wiki and source code
instead.

Some basic information about the project has been retained on the page
to give visitors an overview of what the project's about rather than
diving straight into the source code or wiki.